tests: refactor filepaths & test data
