+++ tests/data/geography_lookup_table.csv
@@ -1,0 +1,4 @@
+,stop_id,stop_lat,stop_lon,LSOA21CD,LSOA21NM,BUA22CD,BUA22NM,LAD23CD,LAD23NM,RGN22CD,RGN22NM,TTWA11CD,TTWA11NM
+0,0100BRP90310,51.44902101682718,-2.5857890312830363,E01033353,Bristol 054B,,,E06000023,"Bristol, City of",E12000009,South West,E30000180,Bristol
+1,3200YND57040,53.98630052493483,-1.52401954306945,E01027733,Harrogate 017F,E63000601,Harrogate,E06000065,North Yorkshire,E12000003,Yorkshire and The Humber,E30000214,Harrogate
+2,490005344S,51.55653865410456,-0.4733171011848141,E01002512,Hillingdon 013C,,,E09000017,Hillingdon,E12000007,London,E30000266,Slough and Heathrow
