+++ tests/data/stop_level_punctuality.csv
@@ -1,0 +1,4 @@
+,stop_id,stop_lat,stop_lon,service_stops,punctuality_rate
+0,0100BRP90310,51.44902101682718,-2.5857890312830363,17,0.17
+1,490005344S,51.55653865410456,-0.4733171011848141,23,0.45
+2,3200YND57040,53.98630052493483,-1.52401954306945,4,1.0
